Create README.md
Personal web site for my training business. I am using Django Python and Javascript. Python and Django are being used currently to pull data and post videos to the main page of clients and myself performing exercises. Javascript is being used along with Google Sheets as a back end to pull sheet data and populate that data onto a page in HTML format. The web page will be responsive to mobile, tablets and PC. The Javascript feature is being worked on for a separate page. As the page grows I will be adding a login and registration page with authorizations and password protections with a Django password recovery system in place